pyproject.toml
[tool.poetry]
name = "django-petvida"
version = "0.1.0"
description = "app para gerenciamento de vacinas e tratamento de uma veterinaria"
authors = ["Heiden <119899341+CarlosHeiden@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.1,<3.11"
django = "^5.2.4"
mysqlclient = "^2.2.7"
djangorestframework = "^3.16.1"
firebase-admin = "^7.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
